+++ cmd/study/duplicate_test.go
@@ -1,0 +1,74 @@
+package study_test
+
+import (
+	"bufio"
+	"bytes"
+	"fmt"
+	"os"
+	"testing"
+
+	"github.com/benmatselby/prolificli/cmd/study"
+	"github.com/benmatselby/prolificli/mock_client"
+	"github.com/benmatselby/prolificli/model"
+	"github.com/golang/mock/gomock"
+)
+
+func TestNewDuplicateCommand(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	client := mock_client.NewMockAPI(ctrl)
+
+	cmd := study.NewDuplicateCommand(client, os.Stdout)
+
+	use := "duplicate"
+	short := "Duplicate an existing study"
+
+	if cmd.Use != use {
+		t.Fatalf("expected use: %s; got %s", use, cmd.Use)
+	}
+
+	if cmd.Short != short {
+		t.Fatalf("expected use: %s; got %s", short, cmd.Short)
+	}
+}
+
+func TestDuplicateStudyCallsClient(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	c := mock_client.NewMockAPI(ctrl)
+
+	studyID := "11223344"
+
+	actualStudy := model.Study{
+		ID:                      "9999999",
+		Name:                    "My first standard sample",
+		InternalName:            "Standard sample",
+		Desc:                    "This is my first standard sample study on the Prolific system.",
+		ExternalStudyURL:        "https://eggs-experriment.com?participant={{%PROLIFIC_PID%}}",
+		TotalAvailablePlaces:    10,
+		EstimatedCompletionTime: 10,
+		MaximumAllowedTime:      10,
+		Reward:                  400,
+		DeviceCompatibility:     []string{"desktop", "tablet", "mobile"},
+	}
+
+	c.
+		EXPECT().
+		DuplicateStudy(gomock.Eq(studyID)).
+		Return(&actualStudy, nil).
+		AnyTimes()
+
+	var b bytes.Buffer
+	writer := bufio.NewWriter(&b)
+
+	cmd := study.NewDuplicateCommand(c, writer)
+	_ = cmd.RunE(cmd, []string{studyID})
+	writer.Flush()
+
+	expected := fmt.Sprintf("%s\n", actualStudy.ID)
+	actual := b.String()
+
+	if actual != expected {
+		t.Fatalf("expected \n'%s'\ngot\n'%s'", expected, actual)
+	}
+}